changed tests to new routing!!
